+++ src/app/game/game-mode/base-game-mode.ts/set-promode-temp-vision-state.ts
@@ -1,4 +1,4 @@
-import { setProModeTempVision } from '../base/utillity/pro-mode-temp-vision';
+import { setProModeTempVision } from '../utillity/pro-mode-temp-vision';
 import { BaseState } from '../state/base-state';
 import { StateData } from '../state/state-data';
