+++ src/app/game/game-mode/mode/base-game-mode.ts
@@ -1,6 +1,5 @@
 import { BaseState } from '../state/base-state';
 import { StateData } from '../state/state-data';
-import { debugPrint } from 'src/app/utils/debug-print';
 
 export class BaseData implements StateData {}
