+++ src/app/game/game-mode/utillity/remove-units.ts
