+++ src/app/game/game-mode/utillity/update-ui.ts
