+++ src/configs/game-settings.ts
@@ -6,7 +6,7 @@ export const CITIES_TO_WIN_RATIO: number = 0.6;
 export const STARTING_INCOME: number = 4;
 
 //This is the duration of a turn in seconds. 60 seconds by default.
-export const TURN_DURATION_IN_SECONDS: number = 60;
+export const TURN_DURATION_IN_SECONDS: number = 10;
 
 //This is the duration of a tick in seconds. 1 second by default.
 export const TICK_DURATION_IN_SECONDS: number = 1;
